Split the parameter validation functions off into... ... their own sub-module to simplify testing and make it simpler to re-use these routines in algorithms ('cause there's likely some duplication of effort over there that I'm probably going to live with for v0.5).
