+++ src/digraph.js
@@ -4,107 +4,12 @@
 // http://en.wikipedia.org/wiki/Directed_graph
 
 var helperFunctions = require('./helper-functions');
+var digraphParams = require('./digraph-in-parameters');
 var digraphImport = require('./digraph-json-import');
 var digraphExport = require('./digraph-json-export');
 
 (function() {
     var __bind = function(method, scope){ return function(){ return method.apply(scope, arguments); }; };
-
-    var verifyVertexReadRequest = function(request_) {
-        var response = { error: null, result: false };
-        var jstype = helperFunctions.JSType(request_);
-        if (jstype !== '[object String]') {
-            response.error = "Invalid value type '" + jstype + "' found vertex read request. Expected '[object String]'.";
-        } else {
-            response.result = true;
-        }
-        return response;
-    };
-
-    var verifyVertexWriteRequest = function(request_) {
-        var response = { error: null, result: false };
-        var inBreakScope = false;
-        while (!inBreakScope) {
-            inBreakScope = true;
-            var jstype = helperFunctions.JSType(request_);
-            if (jstype !== '[object Object]') {
-                response.error = "Invalid value type '" + jstype + "' found when expecting a vertex write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.u);
-            if (jstype !== '[object String]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for vertex ID string property 'u' in vertex write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.p);
-            if (jstype === '[object Function]') {
-                response.error = "Invalid value type '" + jstype + " found while inspecting vertex property 'p' in vertex write request object. Must be serializable to JSON!";
-                break;
-            }
-            response.result = true;
-        }
-        return response;
-    };
-
-    var verifyEdgeReadRequest = function(request_) {
-        var response = { error: null, result: false };
-        var inBreakScope = false;
-        while (!inBreakScope) {
-            inBreakScope = true;
-            var jstype = helperFunctions.JSType(request_);
-            if (jstype !== '[object Object]') {
-                response.error = "Invalid value type '" + jstype + "' found when expecting edge read request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.u);
-            if (jstype !== '[object String]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for vertex ID string property 'u' in edge read request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.v);
-            if (jstype !== '[object String]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for vertex ID string property 'v' in edge read request object.";
-                break;
-            }
-            response.result = true;
-        }
-        return response;
-    };
-
-    var verifyEdgeWriteRequest = function(request_) {
-        var response = { error: null, result: false };
-        var inBreakScope = false;
-        while (!inBreakScope) {
-            inBreakScope = true;
-            var jstype = helperFunctions.JSType(request_);
-            if (jstype !== '[object Object]') {
-                response.error = "Invalid value type '" + jstype + "' found when expecting edge write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.e);
-            if (jstype !== '[object Object]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for edge descriptor object 'e' in edge write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.e.u);
-            if (jstype !== '[object String]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for vertex ID string property 'e.u' in edge write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.e.v);
-            if (jstype !== '[object String]') {
-                response.error = "Invalid value type '" + jstype + "' found looking for vertex ID string property 'e.v' in edge write request object.";
-                break;
-            }
-            jstype = helperFunctions.JSType(request_.p);
-            if (jstype === '[object Function]') {
-                response.error = "Invalid value type '" + jstype + "' found while insecting edge property 'p' in edge write request object. Must be serializable to JSON!";
-            }
-            response.result = true;
-        }
-        return response;
-    };
-    
 
     var DirectedGraph = (function() {
         function DirectedGraph(jsonOrObject_) {
@@ -201,7 +106,7 @@ var digraphExport = require('./digraph-json-export');
             var inBreakScope = false;
             while (!inBreakScope) {
                 inBreakScope = true;
-                var innerResponse = verifyVertexWriteRequest(request_);
+                var innerResponse = digraphParams.verifyVertexWriteRequest(request_);
                 if (innerResponse.error) {
                     errors.unshift(innerResponse.error);
                     break;
@@ -228,7 +133,7 @@ var digraphExport = require('./digraph-json-export');
         };
 
         DirectedGraph.prototype.isVertex = function (vertexId_) {
-            var innerResponse = verifyVertexReadRequest(vertexId_);
+            var innerResponse = digraphParams.verifyVertexReadRequest(vertexId_);
             if (innerResponse.error) {
                 return false;
             }
@@ -237,7 +142,7 @@ var digraphExport = require('./digraph-json-export');
         };
             
         DirectedGraph.prototype.removeVertex = function (vertexId_) {
-            var innerResponse = verifyVertexReadRequest(vertexId_);
+            var innerResponse = digraphParams.verifyVertexReadRequest(vertexId_);
             if (innerResponse.error) {
                 return false;
             }
@@ -274,7 +179,7 @@ var digraphExport = require('./digraph-json-export');
             var inBreakScope = false;
             while (!inBreakScope) {
                 inBreakScope = true;
-                var innerResponse = verifyEdgeWriteRequest(request_);
+                var innerResponse = digraphParams.verifyEdgeWriteRequest(request_);
                 if (innerResponse.error) {
                     errors.unshift(innerResponse.error);
                     break;
@@ -328,7 +233,7 @@ var digraphExport = require('./digraph-json-export');
             var inBreakScope = false;
             while (!inBreakScope) {
                 inBreakScope = true;
-                var innerResponse = verifyEdgeReadRequest(request_);
+                var innerResponse = digraphParams.verifyEdgeReadRequest(request_);
                 if (innerResponse.error) {
                     errors.unshift(innerResponse.error);
                     break;
@@ -379,7 +284,7 @@ var digraphExport = require('./digraph-json-export');
             var inBreakScope = false;
             while (!inBreakScope) {
                 inBreakScope = true;
-                if (verifyEdgeReadRequest(request_).error) {
+                if (digraphParams.verifyEdgeReadRequest(request_).error) {
                     break;
                 }
                 var vertexU = this._private.vertexMap[request_.u];
@@ -480,7 +385,7 @@ var digraphExport = require('./digraph-json-export');
             var inBreakScope = false;
             while (!inBreakScope) {
                 inBreakScope = true;
-                if (verifyEdgeReadRequest(request_).error) {
+                if (digraphParams.verifyEdgeReadRequest(request_).error) {
                     break;
                 }
                 var vertexU = this._private.vertexMap[request_.u];
@@ -555,6 +460,7 @@ var digraphExport = require('./digraph-json-export');
           response.error string returned by DirectedGraph's data import subroutine.
         */
         DirectedGraph: DirectedGraph
+
     };
 
 }).call(this);
